backend/requirements.txt
sqlalchemy
pydantic
fastapi
uvicorn
requests
psycopg2-binary
python-dotenv